library/requirements.txt
Django==2.2.14
django-smtp-ssl==1.0
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
